+++ packages/plugin-prefetch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @spoosh/plugin-prefetch
 
+## 0.3.3
+
+- Changed `PluginContext.path` from `string[]` to `string` for simpler plugin API
+
 ## 0.3.2
 
 - Update documentation URLs to new `/docs/{framework}` format