+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @spoosh/test-utils
 
+## 0.1.6
+
+- Changed `PluginContext.path` from `string[]` to `string` for simpler plugin API
+
 ## 0.1.5
 
 ### Patch Changes